moved all but main to python_timer_stuff/__init__.py
